8 puzzle problem
Included some reports written for the class. 

The program only has one file.

Run the program and follow the instructions. 

The program should prompt the user to choose which heuristic function to use.
The program should also prompt the user for the method of input. A straight line input 
is if the user inputs the numbers 123456780. 

The time and average node should be displayed at the end of the puzzle solving for the 
multiple puzzle choice. This was for the 100 cases. This mode will only work for 100 cases no less, no more.